% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_case_death_counts.R
\name{add_case_death_counts}
\alias{add_case_death_counts}
\title{A function to calculate the death case counts and number of days since a
specific number of cases or deaths are reported.}
\usage{
add_case_death_counts(case_count_n = 0, death_count_m = 0)
}
\arguments{
\item{case_count_n}{Number of cases from which it starts a counter of number
of days passed.}

\item{death_count_m}{Number of deaths from which it starts a counter of
number of days passed.}
}
\value{
Returns the data frame with daily data of cases and deaths for each
  country.
}
\description{
This function downloads the latest corona wordwide case and death data as per
the data format of European Centre for Disease Prevention and Control. It
calculates the number of cumulative cases, cumulative deaths and number of
days since the first case, nth case, first death and mth death are reported
for each country and add them as additional columns in data frame.
}
\examples{
covid_world_data_today <- add_case_death_counts(200,20)
}
\keyword{add_case_death_counts}
