% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidanalyzr.R
\docType{package}
\name{covidanalyzr}
\alias{covidanalyzr}
\title{covidanalyzr: An implementation of functions to analyze and visualise
Covid-19 Data}
\description{
The covidanalyzr package provides important functions:
add_case_death_counts, print_case_graph, print_case_tiles , print_death_graph
}
\section{covidanalyzr functions}{

add_case_death_counts
print_case_graph
print_case_tiles
print_death_graph
}

