% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_case_graph.R
\name{print_case_graph}
\alias{print_case_graph}
\title{A function to visualize a line plot of growth of confirmed corona virus infections after first n cases are confirmed, grouped by country.}
\usage{
print_case_graph(countries, n = 0)
}
\arguments{
\item{countries}{A vector of all country names that the corona case count need to be plotted.}

\item{n}{number of confirmed covid cases beyond which the number of days for the plot will begin.}
}
\description{
This function downloads ordwide case and death data as per the data format of European Centre for Disease Prevention and Control. Based on countries given as input parameter it plots a line graph of how the confirmed cases are progressing by day.
}
\examples{
print_case_graph(c("India", "United"States"of_Ameria"), 200)
}
\keyword{print_case_graph}
